Added gravestone recipe to last page of soul campfire research

+++ src/main/java/thaumicindustries2core/asm/ClaimMapTransformer.java
@@ -244,7 +244,7 @@ public class ClaimMapTransformer implements IClassTransformer {
         }
         if (n.getOpcode() == Opcodes.LDC && n instanceof LdcInsnNode) {
             Object cst = ((LdcInsnNode) n).cst;
-            return cst instanceof Float && ((Float) cst).floatValue() == value;
+            return cst instanceof Float && (Float) cst == value;
         }
         return false;
     }
